refactor(mcp): extract visibility helpers to tools/visibility.rs
Move ensure_element_in_view (260 lines) and ensure_visible_and_apply_highlight
(45 lines) to dedicated module with proper organization:
- Extract nested helper functions to module-level
- Add platform-specific viewport bounds helpers
- Net: -310 lines from server.rs (8556 → 6712, -21.5% total reduction)
